backend/django/requirements.txt
# Django and REST Framework
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10
duckdb==1.3.2

# Data Processing
pandas==2.3.2
pyarrow==21.0.0
numpy==2.3.1
openpyxl==3.1.2

# Task Queue
celery==5.5.3
redis==6.4.0

# Development
python-decouple==3.8
django-extensions==3.2.3

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Monitoring
django-health-check==3.18.1

# Security & Authentication
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
django-validators==1.0.1
cryptography==43.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.9

# Development Tools
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Additional dependencies for better compatibility
Pillow==10.4.0
python-dateutil==2.9.0
